[phi]mv GetKernelTypeForVar to onednn
